Add a client endpoint example, and tweak the output for echo samples
The echo samples were somewhat unclear about what output was coming from (or being echoed to) where,
so I've tried to clarify that with changes to the message formatting to show which messages are being
forwarded from other sessions vs. being echoed back to the originating session.

Tweaks were made to the javascript (to allow a client to be started), and to the html (for instructions).

An empty manifest was checked in because maven kept generating it anyway.